- add StandardQuanticPipeline - Create a utils file for containing dimension reduction algorithms
