+++ examples/ERP/compare_dim_red.py
@@ -13,22 +13,24 @@ It is compared to the classical SVM.
 # Modified from plot_classify_EEG_tangentspace.py
 # License: BSD (3-clause)
 
-import numpy as np
+
 from pyriemann.estimation import XdawnCovariances
 from pyriemann.tangentspace import TangentSpace
-from pyriemann_qiskit.classification import QuanticSVM
+from pyriemann_qiskit.classification import QuanticSVM, QuanticVQC, StandardQuanticPipeline
+from pyriemann_qiskit.utils.filtering import NaivePair, NaiveImpair
 from mne import io, read_events, pick_types, Epochs
 from mne.datasets import sample
 from sklearn.pipeline import make_pipeline
-from sklearn.model_selection import train_test_split
+
 from sklearn.metrics import confusion_matrix, ConfusionMatrixDisplay
-from sklearn.base import TransformerMixin
-from sklearn.svm import SVC
+
+
 from sklearn.decomposition import PCA
 from matplotlib import pyplot as plt
 from sklearn.model_selection import StratifiedKFold, cross_val_score, cross_val_predict
 from sklearn.metrics import balanced_accuracy_score
-import itertools
+
+
 
 
 print(__doc__)
@@ -66,6 +68,8 @@ epochs = Epochs(
     verbose=False,
 )
 
+epochs = epochs[:20]
+
 X = epochs.get_data()
 y = epochs.events[:, -1]
 
@@ -78,91 +82,7 @@ y = epochs.events[:, -1]
 # Thus we reduce elements number by using restrictive spatial filtering
 sf = XdawnCovariances(nfilter=2)
 
-class NoFilter(TransformerMixin):
-    def fit(self, X, y=None):
-        return self
 
-    def transform(self, X, y=None):
-        return X
-
-class NaivePair(TransformerMixin):
-    def fit(self, X, y=None):
-        return self
-
-    def transform(self, X, y=None):
-        return X[:, ::2]
-
-class NaiveImpair(TransformerMixin):
-    def fit(self, X, y=None):
-        return self
-
-    def transform(self, X, y=None):
-        return X[:, 1::2]
-
-class Closest(TransformerMixin):
-    def __init__(self, ndim=None):
-        self.ndim = ndim
-        self.closest_subset_to_vector = None
-
-    def fit(self, X, y=None):
-        n_ts = X.shape[1]
-        if self.ndim is None:
-            self.ndim = n_ts // 2
-        indices = range(n_ts)
-        ret = {}
-        print(indices, self.ndim)
-        for v in X:
-            for subset in itertools.combinations(indices, self.ndim):
-                not_indices = [i for i in indices if i not in subset]
-                sub_v = v.copy()
-                sub_v[not_indices] = 0
-                dist = np.linalg.norm(v - sub_v)
-                key = ''.join(str(val) for val in subset)
-                if key in ret:
-                    ret[key] = ret[key] + dist
-                else:
-                    ret[key] = dist
-        self.closest_subset_to_vector = min(ret, key=ret.get)
-        print("Fit finished")
-        return self
-
-    def transform(self, X, y=None):
-        subset = [int(i) for i in self.closest_subset_to_vector]
-        ret = X[:, subset]
-        return ret
-
-# ...and dividing the number of remaining elements by two
-class Preclassif(TransformerMixin):
-    def __init__(self, ndim=None):
-        self.ndim = ndim
-        self.closest_subset_to_vector = None
-
-    def fit(self, X, y=None):
-        n_ts = X.shape[1]
-        if self.ndim is None:
-            self.ndim = n_ts // 2
-        indices = range(n_ts)
-        ret = {}
-        for subset in itertools.combinations(indices, self.ndim):
-            sub_vectors = None
-            for v in X:
-                if sub_vectors is None:
-                    sub_vectors = np.atleast_2d([v[list(subset)]])
-                else:
-                    sub_vectors = np.append(sub_vectors, [v[list(subset)]], axis=0)
-            svc = SVC()
-
-            svc.fit(sub_vectors, y)
-            key = ''.join(str(val) for val in subset)
-            ret[key] = svc.score(sub_vectors, y)
-        self.closest_subset_to_vector = max(ret, key=ret.get)
-        print("Fit finished")
-        return self
-
-    def transform(self, X, y=None):
-        subset = [int(i) for i in self.closest_subset_to_vector]
-        ret = X[:, subset]
-        return ret
 
 
 # Projecting correlation matrices into the tangent space
@@ -174,7 +94,7 @@ tg = TangentSpace()
 
 
 # filters = [Closest(ndim=5), Preclassif(ndim=5), NaivePair(), NaiveImpair(), PCA(n_components=5), NoFilter()]
-filters = [NaivePair(), PCA(n_components=5)]
+filters = [NaivePair(), NaiveImpair()]
 n_filters = len(filters)
 
 f, axes = plt.subplots(2, n_filters, sharey='row')
@@ -192,9 +112,10 @@ for quantum in [False, True]:
         print("running filter " + filter_name)
         cv = StratifiedKFold(n_splits=5, shuffle=False)
         
-        clf = make_pipeline(sf, tg, filter, QuanticSVM(quantum=quantum))
+        # pipe = make_pipeline(sf, tg, filter, QuanticSVM(quantum=quantum))
+        pipe = StandardQuanticPipeline(dim_red=filter, quantum_clf=QuanticSVM(quantum=quantum))
         # score = cross_val_score(clf, X, y, cv=cv, scoring='balanced_accuracy').mean()
-        y_pred = cross_val_predict(clf, X, y, cv=cv)
+        y_pred = cross_val_predict(pipe, X, y, cv=cv)
 
         score = balanced_accuracy_score(y_pred, y)
         # Printing the results
